% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asyncr.R
\name{asyncr_attrib}
\alias{asyncr_attrib}
\title{FUNCTION_TITLE}
\usage{
asyncr_attrib(test_driver, e, attrib, maxiter = 20)
}
\arguments{
\item{test_driver}{PARAM_DESCRIPTION}

\item{e}{PARAM_DESCRIPTION}

\item{attrib}{PARAM_DESCRIPTION}

\item{maxiter}{PARAM_DESCRIPTION, Default: 20}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
