% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_busy_time}
\alias{get_busy_time}
\alias{reset_busy_time}
\title{Shiny busy loggers}
\usage{
get_busy_time(obj, history = FALSE)

reset_busy_time(obj)
}
\arguments{
\item{obj}{reactor object}

\item{history}{logical, return all the logged times,
or only the last logged time? Default: FALSE}
}
\value{
numeric
}
\description{
Query/Reset shiny busy loggers
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
Other reactor: 
\code{\link{init_reactor}()},
\code{\link{kill_app}()},
\code{\link{navigate_to_app}()},
\code{\link{reactor_path}()},
\code{\link{read_reactor}()},
\code{\link{reset_counter}()},
\code{\link{start_reactor}()},
\code{\link{wait_for_shiny}()}
}
\concept{reactor}
