% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{chrome_options}
\alias{chrome_options}
\title{FUNCTION_TITLE}
\usage{
chrome_options(download_path = tempdir())
}
\arguments{
\item{download_path}{PARAM_DESCRIPTION, Default: tempdir()}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
