% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{chrome_version}
\alias{chrome_version}
\title{FUNCTION_TITLE}
\usage{
chrome_version()
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[pagedown]{find_chrome}}
}
