% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actions.R
\name{click_id}
\alias{click_id}
\title{Click an element}
\usage{
click_id(obj, id)
}
\arguments{
\item{obj}{reactor object}

\item{id}{character, id of the element}
}
\value{
reactor object
}
\description{
Send a click command to an element in the application.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[glue]{glue}}

Other actions: 
\code{\link{execute}()},
\code{\link{query}()},
\code{\link{set_id_value}()}
}
\concept{actions}
