% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{driver}
\alias{driver}
\title{FUNCTION_TITLE}
\usage{
driver(test_path = tempdir(), check = TRUE, verbose = FALSE,
  port = 4567L, c_opts = chrome_options(download_path = test_path))
}
\arguments{
\item{test_path}{PARAM_DESCRIPTION, Default: tempdir()}

\item{check}{PARAM_DESCRIPTION, Default: TRUE}

\item{verbose}{PARAM_DESCRIPTION, Default: FALSE}

\item{port}{PARAM_DESCRIPTION, Default: 4567}

\item{c_opts}{PARAM_DESCRIPTION, Default: chrome_options(download_path = test_path)}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[RSelenium]{rsDriver}}
}
