% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drivers.R
\name{chrome_options}
\alias{chrome_options}
\alias{firefox_options}
\title{Chrome Driver Options}
\usage{
chrome_options(
  cargs = c("--disable-gpu", "--window-size=1280,800"),
  headless = TRUE,
  ...,
  download_path = tempdir()
)

firefox_options(
  cargs = c("--width=1280", "--height=800", "--memory 1024mb"),
  headless = TRUE,
  ...,
  download_path = tempdir()
)
}
\arguments{
\item{cargs}{character arguments to pass to driver.}

\item{headless}{logical, is the driver run in headless state. Default: TRUE}

\item{\dots}{additional preferences to add to driver.}

\item{download_path}{character, Path to save files to. Default: tempdir()}
}
\value{
list
}
\description{
Arguments and preferences to pass to chrome driver
}
\details{
By default the preferences are set to allow for local dowloading to a user defined path.
\itemize{
\item chrome:
\itemize{
\item "profile.default_content_settings.popups" = 0L
\item download.prompt_for_download" = FALSE
\item download.directory_upgrade" = TRUE
\item safebrowsing.enabled" = TRUE
\item download.default_directory" = download_path
}
\item firefox:
\itemize{
\item 'browser.download.folderList' = 2
\item 'browser.download.lastDir' = download_path
\item 'browser.download.dir' = download_path
}
}
}
\examples{
chrome_options()
firefox_options()

}
