% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drivers.R
\name{chrome_version}
\alias{chrome_version}
\alias{gecko_version}
\title{Driver Version}
\usage{
chrome_version()

gecko_version()
}
\value{
character
}
\description{
Retrieves chrome driver version installed.
}
\examples{
chrome_version()
gecko_version()
}
\seealso{
\code{\link[pagedown]{find_chrome}}

Other driver: 
\code{\link{chrome_options}()},
\code{\link{firefox_driver}()},
\code{\link{set_chrome_driver}()},
\code{\link{set_firefox_driver}()},
\code{\link{set_timeout}()}

Other driver: 
\code{\link{chrome_options}()},
\code{\link{firefox_driver}()},
\code{\link{set_chrome_driver}()},
\code{\link{set_firefox_driver}()},
\code{\link{set_timeout}()}
}
\concept{driver}
