% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actions.R
\name{execute}
\alias{execute}
\title{Execute Command in Application}
\usage{
execute(obj, expr)
}
\arguments{
\item{obj}{reactor object}

\item{expr}{character, Javascript command}
}
\value{
reactor object
}
\description{
Execute JavaScript function in application.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
Other actions: 
\code{\link{click_id}()},
\code{\link{query_style_id}()},
\code{\link{query}()},
\code{\link{set_id_value}()}
}
\concept{actions}
