% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actions.R
\name{execute_script}
\alias{execute_script}
\title{FUNCTION_TITLE}
\usage{
execute_script(obj, expr)
}
\arguments{
\item{obj}{PARAM_DESCRIPTION}

\item{expr}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
