% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect.R
\name{expect_busy_time}
\alias{expect_busy_time}
\title{Expectation: How long is shiny busy?}
\usage{
expect_busy_time(object, max_time, history = FALSE)
}
\arguments{
\item{object}{reactor object}

\item{max_time}{numeric, maximum expected time}

\item{history}{logical, calculate the full history of time
or only the last operation, Default: FALSE}
}
\value{
invisible reactor object
}
\description{
Test the expectation that shiny was busy for an interval of time.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[testthat]{expect}}

Other testing: 
\code{\link{expect_reactivity}()},
\code{\link{test_app}()}
}
\concept{testing}
