% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect.R
\name{expect_count}
\alias{expect_count}
\title{FUNCTION_TITLE}
\usage{
expect_count(counter, tag, exptected)
}
\arguments{
\item{counter}{PARAM_DESCRIPTION}

\item{tag}{PARAM_DESCRIPTION}

\item{exptected}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[testthat]{equality-expectations}}
}
