% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect.R
\name{expect_reactivity}
\alias{expect_reactivity}
\title{Expectation: is the counter equal to a value?}
\usage{
expect_reactivity(object, tag, count)
}
\arguments{
\item{object}{whereami counter object read from whereami.json}

\item{tag}{character, tag name of the element in the counter object to query}

\item{count}{numeric, expected value}
}
\value{
invisible result
}
\description{
Compares the maximum counter value of a tag in a \emph{whereami} counter
object with a value
}
\examples{
if(interactive()){

txt <- "
whereami::cat_where(whereami::whereami(tag = 'reactor test'))
"

whereami::set_whereami_log(tempdir())

tf <- tempfile(fileext = '.R')

cat(txt,file = tf)

source(tf)

counter <- read_reactor(path = tempdir())

expect_reactivity(counter,'reactor test',1)

# clean up temp files

whereami::counter_reset()
unlink(tf)
unlink(reactor_dir, recursive = TRUE, force = TRUE)

}

}
