% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{export_shiny}
\alias{export_shiny}
\title{Exports the notebook and an app.R file for launching the notebook as a shiny application}
\usage{
export_shiny(notebook, directory)
}
\arguments{
\item{notebook}{Reactor notebook}

\item{directory}{directory to place notebook and app.R file}
}
\description{
Exports the notebook and an app.R file for launching the notebook as a shiny application
}
\details{
This function creates two files in the supplied directory: notebook.Rmd, which contains the supplied Reactor notebook, and app.R,
which is a short file that launches the notebook as a Shiny application. The resulting folder functions as a Shiny application
suitable for deployment on e.g. shinyapps.io.
}
