% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_timeout.R
\name{file_timeout}
\alias{file_timeout}
\title{Timer for waiting for File}
\usage{
file_timeout(path, maxiter = 20)
}
\arguments{
\item{path}{character, Path to look for file}

\item{maxiter}{numeric, maximum iteration to wait, Default: 20}
}
\value{
logical
}
\description{
Wait for file to be creating in a local folder
}
