% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{init_reactor}
\alias{init_reactor}
\title{FUNCTION_TITLE}
\usage{
init_reactor()
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
