% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asyncr.R
\name{is_identical}
\alias{is_identical}
\title{Tests objects for Exact Quality}
\usage{
is_identical(x, y, ...)
}
\arguments{
\item{x}{any \R objects.}

\item{y}{any \R objects.}

\item{\dots}{arguments passed to \link[base:identical]{identical}}
}
\value{
value of \strong{y} if x and y are identical, else NULL
}
\description{
Wrapper on \link[base:identical]{identical} that will return the
value of y if TRUE, else NULL. This function is useful with \link[=wait]{wait}/\link[=then]{then}
}
\seealso{
\link[=wait]{wait}, \link[=then]{then}, \link[base:identical]{identical}
}
