% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{kill_app}
\alias{kill_app}
\title{FUNCTION_TITLE}
\usage{
kill_app(obj, processx_cleanup = TRUE)
}
\arguments{
\item{obj}{PARAM_DESCRIPTION}

\item{processx_cleanup}{logical, cleanup the side effects created by processx. Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
