% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Server.R
\name{launch_reactive_notebook}
\alias{launch_reactive_notebook}
\title{Launches a ReactiveNotebook server}
\usage{
launch_reactive_notebook(notebook)
}
\arguments{
\item{notebook}{ReactiveNotebook to view in the web-based notebook editor}
}
\description{
Launches a ReactiveNotebook server
}
\examples{
\dontrun{
library(ReactiveNotebook)
notebook <- ReactiveNotebook$new()

server <- launch_reactive_notebook(notebook)

stop_reactive_notebook(server)
}

}
