% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_args.R
\name{manual_args}
\alias{manual_args}
\title{FUNCTION_TITLE}
\usage{
manual_args(expr, test_path = tempdir())
}
\arguments{
\item{expr}{PARAM_DESCRIPTION}

\item{test_path}{PARAM_DESCRIPTION, Default: tempdir()}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
