% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{print.reactor}
\alias{print.reactor}
\title{FUNCTION_TITLE}
\usage{
\method{print}{reactor}(x, ...)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[yaml]{as.yaml}}
}
