% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asyncr.R
\name{wait}
\alias{wait}
\alias{then}
\title{FUNCTION_TITLE}
\usage{
wait(test_driver, expr, maxiter = 20)

then(elem, expr, test_driver, maxiter = 20)
}
\arguments{
\item{test_driver}{PARAM_DESCRIPTION}

\item{expr}{PARAM_DESCRIPTION}

\item{maxiter}{PARAM_DESCRIPTION, Default: 20}

\item{elem}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
