% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asyncr.R
\name{then2}
\alias{then2}
\title{FUNCTION_TITLE}
\usage{
then2(elem, elem2, expr, test_driver, maxiter = 20)
}
\arguments{
\item{elem}{PARAM_DESCRIPTION}

\item{elem2}{PARAM_DESCRIPTION}

\item{expr}{PARAM_DESCRIPTION}

\item{test_driver}{PARAM_DESCRIPTION}

\item{maxiter}{PARAM_DESCRIPTION, Default: 20}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
