% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actions.R
\name{query}
\alias{query}
\title{FUNCTION_TITLE}
\usage{
query(obj, command)
}
\arguments{
\item{obj}{PARAM_DESCRIPTION}

\item{command}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[glue]{glue}}
}
