% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actions.R
\name{query_style_id}
\alias{query_style_id}
\title{Query CSS Style}
\usage{
query_style_id(obj, id, style, flatten = FALSE)
}
\arguments{
\item{obj}{reactor object}

\item{id}{character, id of the element}

\item{style}{name of the style element}

\item{flatten}{logical, flatten the output list. Default: FALSE}
}
\description{
Query the style value of an element by id
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[glue]{glue}}

Other actions: 
\code{\link{click_id}()},
\code{\link{execute}()},
\code{\link{query}()},
\code{\link{set_id_value}()}
}
\concept{actions}
