% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReactiveNotebook.R
\docType{data}
\name{ReactiveNotebook}
\alias{ReactiveNotebook}
\title{ReactiveNotebook R6 class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ReactiveNotebook
}
\description{
ReactiveNotebook R6 class
}
\keyword{datasets}
