% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{reactor_example}
\alias{reactor_example}
\title{Load a Reactor example notebook}
\usage{
reactor_example(name)
}
\arguments{
\item{name}{example notebook filename}
}
\description{
Load a Reactor example notebook
}
\seealso{
reactor_example_list
}
