% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_files.R
\name{reactor_path}
\alias{reactor_path}
\title{Construct Path to reactor directory}
\usage{
reactor_path(x = tempdir(), ...)
}
\arguments{
\item{x}{character, root of the path, Default: tempdir()}

\item{...}{additional arguments to pass to file.path}
}
\value{
character
}
\description{
Consructs the path to reactor temporary directory.
}
\examples{
reactor_path()
reactor_path(tempdir(),'internal')
}
\seealso{
Other reactor: 
\code{\link{get_busy_time}()},
\code{\link{init_reactor}()},
\code{\link{kill_app}()},
\code{\link{navigate_to_app}()},
\code{\link{read_reactor}()},
\code{\link{reset_counter}()},
\code{\link{start_reactor}()},
\code{\link{wait_for_shiny}()}
}
\concept{reactor}
