% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReactorNotebook.R
\name{ReactorNotebook}
\alias{ReactorNotebook}
\title{Reactor Notebook R6 class}
\description{
Reactor Notebook R6 class

Reactor Notebook R6 class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ReactorNotebook$new()}}
\item \href{#method-run_in_env}{\code{ReactorNotebook$run_in_env()}}
\item \href{#method-move}{\code{ReactorNotebook$move()}}
\item \href{#method-delete_cell}{\code{ReactorNotebook$delete_cell()}}
\item \href{#method-run_all}{\code{ReactorNotebook$run_all()}}
\item \href{#method-run_cell}{\code{ReactorNotebook$run_cell()}}
\item \href{#method-update_from_view}{\code{ReactorNotebook$update_from_view()}}
\item \href{#method-update_size}{\code{ReactorNotebook$update_size()}}
\item \href{#method-update_open}{\code{ReactorNotebook$update_open()}}
\item \href{#method-propagate_updates}{\code{ReactorNotebook$propagate_updates()}}
\item \href{#method-data_frame}{\code{ReactorNotebook$data_frame()}}
\item \href{#method-get_graph}{\code{ReactorNotebook$get_graph()}}
\item \href{#method-print}{\code{ReactorNotebook$print()}}
\item \href{#method-save}{\code{ReactorNotebook$save()}}
\item \href{#method-load_}{\code{ReactorNotebook$load_()}}
\item \href{#method-export}{\code{ReactorNotebook$export()}}
\item \href{#method-export_shiny}{\code{ReactorNotebook$export_shiny()}}
\item \href{#method-clone}{\code{ReactorNotebook$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactorNotebook$new(static_dir = tempdir())}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run_in_env"></a>}}
\if{latex}{\out{\hypertarget{method-run_in_env}{}}}
\subsection{Method \code{run_in_env()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactorNotebook$run_in_env(code)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-move"></a>}}
\if{latex}{\out{\hypertarget{method-move}{}}}
\subsection{Method \code{move()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactorNotebook$move(source, destination)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_cell"></a>}}
\if{latex}{\out{\hypertarget{method-delete_cell}{}}}
\subsection{Method \code{delete_cell()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactorNotebook$delete_cell(cell)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run_all"></a>}}
\if{latex}{\out{\hypertarget{method-run_all}{}}}
\subsection{Method \code{run_all()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactorNotebook$run_all()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run_cell"></a>}}
\if{latex}{\out{\hypertarget{method-run_cell}{}}}
\subsection{Method \code{run_cell()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactorNotebook$run_cell(cell, update = TRUE, capturePlots = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_from_view"></a>}}
\if{latex}{\out{\hypertarget{method-update_from_view}{}}}
\subsection{Method \code{update_from_view()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactorNotebook$update_from_view(cell, value, capturePlots = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_size"></a>}}
\if{latex}{\out{\hypertarget{method-update_size}{}}}
\subsection{Method \code{update_size()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactorNotebook$update_size(cell, value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_open"></a>}}
\if{latex}{\out{\hypertarget{method-update_open}{}}}
\subsection{Method \code{update_open()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactorNotebook$update_open(cell, value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-propagate_updates"></a>}}
\if{latex}{\out{\hypertarget{method-propagate_updates}{}}}
\subsection{Method \code{propagate_updates()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactorNotebook$propagate_updates(cell, capturePlots = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-data_frame"></a>}}
\if{latex}{\out{\hypertarget{method-data_frame}{}}}
\subsection{Method \code{data_frame()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactorNotebook$data_frame()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_graph"></a>}}
\if{latex}{\out{\hypertarget{method-get_graph}{}}}
\subsection{Method \code{get_graph()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactorNotebook$get_graph()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactorNotebook$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-save"></a>}}
\if{latex}{\out{\hypertarget{method-save}{}}}
\subsection{Method \code{save()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactorNotebook$save(file, rds = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-load_"></a>}}
\if{latex}{\out{\hypertarget{method-load_}{}}}
\subsection{Method \code{load_()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactorNotebook$load_(file)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-export"></a>}}
\if{latex}{\out{\hypertarget{method-export}{}}}
\subsection{Method \code{export()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactorNotebook$export()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-export_shiny"></a>}}
\if{latex}{\out{\hypertarget{method-export_shiny}{}}}
\subsection{Method \code{export_shiny()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactorNotebook$export_shiny()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactorNotebook$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
