% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_files.R
\name{read_stderr}
\alias{read_stderr}
\alias{read_json}
\alias{read_stdout}
\title{FUNCTION_TITLE}
\usage{
read_stderr(path = tempdir())

read_json(path = tempdir())

read_stdout(path = tempdir())
}
\arguments{
\item{path}{PARAM_DESCRIPTION, Default: tempdir()}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
