% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_files.R
\name{reset_counter}
\alias{reset_counter}
\title{FUNCTION_TITLE}
\usage{
reset_counter(obj)
}
\arguments{
\item{obj}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
