% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_args.R
\name{runApp_args}
\alias{runApp_args}
\title{FUNCTION_TITLE}
\usage{
runApp_args(appDir = getwd(), test_ip = 6012, test_path = tempdir(),
  host = getOption("shiny.host", "127.0.0.1"))
}
\arguments{
\item{appDir}{PARAM_DESCRIPTION, Default: getwd()}

\item{test_ip}{PARAM_DESCRIPTION, Default: 6012}

\item{test_path}{PARAM_DESCRIPTION, Default: tempdir()}

\item{host}{PARAM_DESCRIPTION, Default: getOption("shiny.host", "127.0.0.1")}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
