% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drivers.R
\name{set_firefox_driver}
\alias{set_firefox_driver}
\title{FUNCTION_TITLE}
\usage{
set_firefox_driver(
  obj,
  test_path = tempdir(),
  verbose = FALSE,
  port = httpuv::randomPort(),
  opts = firefox_options(download_path = test_path),
  ...
)
}
\arguments{
\item{obj}{PARAM_DESCRIPTION}

\item{test_path}{PARAM_DESCRIPTION, Default: tempdir()}

\item{verbose}{PARAM_DESCRIPTION, Default: FALSE}

\item{port}{PARAM_DESCRIPTION, Default: httpuv::randomPort()}

\item{opts}{PARAM_DESCRIPTION, Default: firefox_options(download_path = test_path)}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
