% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actions.R
\name{set_id_value}
\alias{set_id_value}
\title{Set the value of an element}
\usage{
set_id_value(obj, id, value)
}
\arguments{
\item{obj}{reactor object}

\item{id}{character, id of the element}

\item{value}{value to set the element to}
}
\value{
reactor object
}
\description{
Set the value of an element by element id.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[glue]{glue}}

Other actions: 
\code{\link{click_id}()},
\code{\link{execute}()},
\code{\link{query_style_id}()},
\code{\link{query}()}
}
\concept{actions}
