% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_args.R
\name{set_runapp_args}
\alias{set_runapp_args}
\title{FUNCTION_TITLE}
\usage{
set_runapp_args(
  obj,
  appDir = getwd(),
  test_port = httpuv::randomPort(),
  test_path = tempdir(),
  test_ip = getOption("shiny.host", "127.0.0.1")
)
}
\arguments{
\item{obj}{PARAM_DESCRIPTION}

\item{appDir}{PARAM_DESCRIPTION, Default: getwd()}

\item{test_port}{PARAM_DESCRIPTION, Default: httpuv::randomPort()}

\item{test_path}{PARAM_DESCRIPTION, Default: tempdir()}

\item{test_ip}{The IPv4 address that the application should listen on.}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
