% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{set_timeout}
\alias{set_timeout}
\title{Set Implicit Timeout}
\usage{
set_timeout(obj, milliseconds = 10000)
}
\arguments{
\item{obj}{reactor object}

\item{milliseconds}{Time interval to wait in milliseconds, Default: 10000}
}
\value{
obj reactor object
}
\description{
Set the implicit timeout for the webdriver.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
Other driver: 
\code{\link{chrome_options}()},
\code{\link{chrome_version}()},
\code{\link{firefox_driver}()},
\code{\link{set_chrome_driver}()},
\code{\link{set_firefox_driver}()}
}
\concept{driver}
