% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_args.R
\name{shinyAppDir_args}
\alias{shinyAppDir_args}
\title{FUNCTION_TITLE}
\usage{
shinyAppDir_args(appDir, test_ip = 6012, test_path = tempdir())
}
\arguments{
\item{appDir}{PARAM_DESCRIPTION}

\item{test_ip}{PARAM_DESCRIPTION, Default: 6012}

\item{test_path}{PARAM_DESCRIPTION, Default: tempdir()}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
