% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{slider}
\alias{slider}
\title{Generate an HTML slider to use as input}
\usage{
slider(min = 0, max = 100, step = 1, value = mean(c(min, max)),
  title = "")
}
\arguments{
\item{min}{minimum}

\item{max}{maximum}

\item{step}{increment}

\item{value}{starting value}

\item{title}{title of slider to display}
}
\description{
Generate an HTML slider to use as input
}
