% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{start_reactor_as_shiny}
\alias{start_reactor_as_shiny}
\title{Run a reactor notebook in Shiny}
\usage{
start_reactor_as_shiny(notebook)
}
\arguments{
\item{notebook}{Reactor notebook

importFrom shiny fluidPage
importFrom glue glue}
}
\description{
Run a reactor notebook in Shiny
}
