% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Server.R
\name{stop_reactive_notebook}
\alias{stop_reactive_notebook}
\title{Stops a ReactiveNotebook server}
\usage{
stop_reactive_notebook(server)
}
\arguments{
\item{server}{ReactiveNotebook server}
}
\description{
Stops a ReactiveNotebook server
}
