% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Server.R
\name{stop_reactor}
\alias{stop_reactor}
\title{Stops a Reactor server}
\usage{
stop_reactor(server)
}
\arguments{
\item{server}{ReactiveNotebook server}
}
\description{
Stops a Reactor server
}
\examples{
\dontrun{
library(ReactorNotebook)
notebook <- ReactorNotebook$new()

server <- start_reactor(notebook)

stop_reactor(server)
}

}
