% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unittest.R
\name{test_app}
\alias{test_app}
\title{Reactor Unit Testing}
\usage{
test_app(path = "tests/testthat", ...)
}
\arguments{
\item{path}{character, path to testing directory, Default: 'tests/testthat'}

\item{\dots}{arguments to pass to \link[testthat:test_dir]{test_dir}}
}
\value{
\link[testthat:test_dir]{test_dir} output
}
\description{
Run unit tests for shiny apps
}
\details{
reactor unit test files are like testthat test files, just the name is \code{reactor-*.R}, instead of
\code{test-*.R}
}
\seealso{
\link[testthat:test_dir]{test_dir}
}
