% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{text}
\alias{text}
\title{Generate an HTML text input}
\usage{
text(value = "", title = "")
}
\arguments{
\item{value}{initial value}

\item{title}{title of input}
}
\description{
Generate an HTML text input
}
