% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{wait_for_shiny}
\alias{wait_for_shiny}
\title{FUNCTION_TITLE}
\usage{
wait_for_shiny(obj, maxiter = 20)
}
\arguments{
\item{obj}{PARAM_DESCRIPTION}

\item{maxiter}{PARAM_DESCRIPTION, Default: 20}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
