% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{checkbox}
\alias{checkbox}
\title{Generate an HTML checkbox input}
\usage{
checkbox(value = FALSE, title = "")
}
\arguments{
\item{value}{initial value (TRUE or FALSE)}

\item{title}{title of input}
}
\description{
Generate an HTML checkbox input
}
\examples{
# Run this in a cell to display a checkbox input
input <- checkbox(value = FALSE, title = "Your title")

# In another cell, retrieve the value of the checkbox:
as.logical(input)

}
