% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drivers.R
\name{firefox_driver}
\alias{firefox_driver}
\alias{chrome_driver}
\title{Firefox and Chrome Drivers}
\usage{
firefox_driver(
  test_path = tempdir(),
  verbose = FALSE,
  port = 4567L,
  opts = firefox_options(download_path = test_path),
  ...
)

chrome_driver(
  test_path = tempdir(),
  verbose = FALSE,
  port = 4567L,
  opts = chrome_options(download_path = test_path),
  ...
)
}
\arguments{
\item{test_path}{path to run the tests, Default: tempdir()}

\item{verbose}{If TRUE, include status messages (if any)}

\item{port}{Port to run on}

\item{opts}{Driver options from \link{firefox_driver} or \link{chrome_driver}}

\item{...}{Additional arguments to pass to \code{\link[RSelenium]{remoteDriver}}}
}
\value{
object of class \link[RSelenium:remoteDriver-class]{remoteDriver}
}
\description{
Drivers for firefox and chrome to run the Shiny application
}
\examples{
\dontrun{
if(interactive()){
 firefox_rs <- firefox_driver()
 firefox_rs$client$navigate('https://www.google.com')
 firefox_rs$client$screenshot(display = TRUE)
 firefox_rs$client$close()
 firefox_rs$server$stop()
 }
}
}
\seealso{
\code{\link[RSelenium]{rsDriver}}
}
