% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect.R
\name{expect_reactivity}
\alias{expect_reactivity}
\title{Expectation: is the counter equal to a value?}
\usage{
expect_reactivity(object, tag, count)
}
\arguments{
\item{object}{a reactor class object or the path to whereami.json}

\item{tag}{character, tag name of the element in the counter object to query}

\item{count}{numeric, expected value}
}
\value{
invisible result
}
\description{
Compares the maximum counter value of a tag in a \emph{whereami} counter
object with a value
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
}
}
}
