% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_timeout.R
\name{file_timeout}
\alias{file_timeout}
\title{FUNCTION_TITLE}
\usage{
file_timeout(path, maxiter = 20)
}
\arguments{
\item{path}{PARAM_DESCRIPTION}

\item{maxiter}{PARAM_DESCRIPTION, Default: 20}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
