% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{kill_app}
\alias{kill_app}
\title{Close reactor}
\usage{
kill_app(obj, processx_cleanup = TRUE)
}
\arguments{
\item{obj}{reactor object}

\item{processx_cleanup}{logical, cleanup the side effects created by reactor. Default: TRUE}
}
\value{
reactor object
}
\description{
Safely close a reactor session.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
Other reactor: 
\code{\link{init_reactor}()},
\code{\link{navigate_to_app}()},
\code{\link{reactor_path}()},
\code{\link{read_reactor}()},
\code{\link{reset_counter}()},
\code{\link{start_reactor}()},
\code{\link{wait_for_shiny}()}
}
\concept{reactor}
