% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Server.R
\name{launch_reactive_notebook}
\alias{launch_reactive_notebook}
\title{Launches a ReactiveNotebook server}
\usage{
launch_reactive_notebook(notebook)
}
\arguments{
\item{notebook}{notebook to edit}
}
\description{
Launches a ReactiveNotebook server
}
