% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asyncr.R
\name{wait}
\alias{wait}
\alias{then}
\title{Reactor Promises}
\usage{
wait(test_driver, expr, maxiter = 20)

then(elem, expr, test_driver, maxiter = 20)
}
\arguments{
\item{test_driver}{RSelenium \link[RSelenium:rsDriver]{driver} object}

\item{expr}{expression to run}

\item{maxiter}{numeric, maximum number of iterations for wait
and then calls, Default: 20}

\item{elem}{element returned by wait function for then call.}
}
\value{
value determined by expr return.
}
\description{
Wait and Then functions used by reactor to interact with applications.
}
