% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asyncr.R
\name{then2}
\alias{then2}
\title{Reactor Then with two elements}
\usage{
then2(elem, elem2, expr, test_driver, maxiter = 20)
}
\arguments{
\item{elem}{element to pass to expr}

\item{elem2}{element to pass to expr}

\item{expr}{expression to run}

\item{test_driver}{RSelenium \link[RSelenium:rsDriver]{driver} object}

\item{maxiter}{numeric, maximum number of iterations to wait in the then2 call, Default: 20}
}
\value{
value determined by expr return.
}
\description{
Mutli element Then function used by reactor to interact with applications.
}
