% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actions.R
\name{query}
\alias{query}
\alias{query_output_id}
\alias{query_input_id}
\alias{query_inputs}
\alias{query_outputs}
\alias{query_input_names}
\alias{query_output_names}
\title{FUNCTION_TITLE}
\usage{
query(obj, command, flatten = FALSE)

query_output_id(obj, id)

query_input_id(obj, id)

query_inputs(obj, include_clientdata = FALSE)

query_outputs(obj, include_clientdata = FALSE)

query_input_names(obj)

query_output_names(obj)
}
\arguments{
\item{obj}{PARAM_DESCRIPTION}

\item{command}{PARAM_DESCRIPTION}

\item{flatten}{logical, flatten the output list. Default: FALSE}

\item{id}{PARAM_DESCRIPTION}

\item{include_clientdata}{logical, include the client data in the output. Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[glue]{glue}}
}
