% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actions.R
\name{query}
\alias{query}
\alias{query_style_id}
\alias{query_output_id}
\alias{query_input_id}
\alias{query_inputs}
\alias{query_outputs}
\alias{query_input_names}
\alias{query_output_names}
\title{Query Command in Application}
\usage{
query(obj, command, flatten = FALSE)

query_style_id(obj, id, style, flatten = FALSE)

query_output_id(obj, id)

query_input_id(obj, id)

query_inputs(obj, include_clientdata = FALSE)

query_outputs(obj, include_clientdata = FALSE)

query_input_names(obj)

query_output_names(obj)
}
\arguments{
\item{obj}{reactor object}

\item{command}{character, JavaScript command}

\item{flatten}{logical, flatten the output list. Default: FALSE}

\item{id}{character, id of the element}

\item{style}{name of the style element}

\item{include_clientdata}{logical, include the client data in the output. Default: FALSE}
}
\value{
value returned by the query
}
\description{
Send a query to the application and return a value.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[glue]{glue}}

Other actions: 
\code{\link{click_id}()},
\code{\link{execute}()},
\code{\link{set_id_value}()}
}
\concept{actions}
