% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_files.R
\name{reactor_path}
\alias{reactor_path}
\title{FUNCTION_TITLE}
\usage{
reactor_path(x = tempdir(), ...)
}
\arguments{
\item{x}{PARAM_DESCRIPTION, Default: tempdir()}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
