% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReactorNotebook.R
\docType{data}
\name{ReactorNotebook}
\alias{ReactorNotebook}
\title{Reactor Notebook R6 class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ReactorNotebook
}
\description{
Reactor Notebook R6 class
}
\keyword{datasets}
