% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_files.R
\name{read_reactor}
\alias{read_reactor}
\alias{read_stdout}
\alias{read_stderr}
\title{Read Child Process Side Effects}
\usage{
read_reactor(x)

read_stdout(x)

read_stderr(x)
}
\arguments{
\item{x}{path to reactor temporary folder or a reactor class
object. Default: tempdir()}
}
\value{
contents of the file
}
\description{
Read files created by reactor in the test folder
}
\seealso{
Other reactor: 
\code{\link{init_reactor}()},
\code{\link{kill_app}()},
\code{\link{navigate_to_app}()},
\code{\link{reactor_path}()},
\code{\link{reset_counter}()},
\code{\link{start_reactor}()},
\code{\link{wait_for_shiny}()}

Other reactor: 
\code{\link{init_reactor}()},
\code{\link{kill_app}()},
\code{\link{navigate_to_app}()},
\code{\link{reactor_path}()},
\code{\link{reset_counter}()},
\code{\link{start_reactor}()},
\code{\link{wait_for_shiny}()}

Other reactor: 
\code{\link{init_reactor}()},
\code{\link{kill_app}()},
\code{\link{navigate_to_app}()},
\code{\link{reactor_path}()},
\code{\link{reset_counter}()},
\code{\link{start_reactor}()},
\code{\link{wait_for_shiny}()}
}
\concept{reactor}
