% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_files.R
\name{read_reactor}
\alias{read_reactor}
\alias{read_stdout}
\alias{read_stderr}
\title{Read Child Process Side Effects}
\usage{
read_reactor(x)

read_stdout(x)

read_stderr(x)
}
\arguments{
\item{x}{path to reactor temporary folder or a reactor class
object. Default: tempdir()}
}
\value{
contents of the file
}
\description{
Read files created by reactor in the test folder
}
