% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_files.R
\name{read_stderr}
\alias{read_stderr}
\alias{read_stdout}
\alias{read_reactor}
\title{Read Child Process Side Effects}
\usage{
read_stderr(path = file.path(tempdir(), "reactor"))

read_stdout(path = file.path(tempdir(), "reactor"))

read_reactor(path = file.path(tempdir(), "reactor"))
}
\arguments{
\item{path}{character, path to test folder, Default: file.path(tempdir(),'reactor')}
}
\value{
contents of the file
}
\description{
Read files created by \link[reactor:test_reactor]{test_reactor} in the test folder
}
