% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_files.R
\name{reset_counter}
\alias{reset_counter}
\title{Reset the whereami counter object}
\usage{
reset_counter(obj)
}
\arguments{
\item{obj}{reactor object}
}
\value{
reactor object
}
\description{
Resets the whereami counter object to an empty json file.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
Other reactor: 
\code{\link{get_busy_time}()},
\code{\link{init_reactor}()},
\code{\link{kill_app}()},
\code{\link{navigate_to_app}()},
\code{\link{reactor_path}()},
\code{\link{read_reactor}()},
\code{\link{start_reactor}()},
\code{\link{wait_for_shiny}()}
}
\concept{reactor}
