% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drivers.R
\name{set_firefox_driver}
\alias{set_firefox_driver}
\title{Attach Firefox Driver to Reactor}
\usage{
set_firefox_driver(
  obj,
  test_path = tempdir(),
  verbose = FALSE,
  port = httpuv::randomPort(),
  opts = firefox_options(download_path = test_path),
  ...
)
}
\arguments{
\item{obj}{reactor object}

\item{test_path}{character, Path the child process will have access to on the master, Default: tempdir()}

\item{verbose}{logical, start the webdriver verbosely. Default: FALSE}

\item{port}{integer, port to run the webdriver on, Default: httpuv::randomPort()}

\item{opts}{named list, options to initialize firefox with. Default: chrome_options(download_path = test_path)}

\item{...}{additional arguments to pass to \code{\link[RSelenium]{rsDriver}}}
}
\value{
reactor object
}
\description{
Attach firefox driver to a reactor object.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
Other driver: 
\code{\link{chrome_options}()},
\code{\link{chrome_version}()},
\code{\link{firefox_driver}()},
\code{\link{set_chrome_driver}()},
\code{\link{set_timeout}()}
}
\concept{driver}
