% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actions.R
\name{set_id_value}
\alias{set_id_value}
\title{FUNCTION_TITLE}
\usage{
set_id_value(obj, id, value)
}
\arguments{
\item{obj}{PARAM_DESCRIPTION}

\item{id}{PARAM_DESCRIPTION}

\item{value}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[glue]{glue}}
}
