% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{start_reactor}
\alias{start_reactor}
\title{Start the reactor}
\usage{
start_reactor(obj, silent = FALSE)
}
\arguments{
\item{obj}{reactor object}

\item{silent}{logical, start silently. Default: FALSE}
}
\value{
reactor object
}
\description{
Using the populated elements of reactor you can start
the child sessions.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[processx]{process}}

Other reactor: 
\code{\link{get_busy_time}()},
\code{\link{init_reactor}()},
\code{\link{kill_app}()},
\code{\link{navigate_to_app}()},
\code{\link{reactor_path}()},
\code{\link{read_reactor}()},
\code{\link{reset_counter}()},
\code{\link{wait_for_shiny}()}
}
\concept{reactor}
