% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{start_reactor}
\alias{start_reactor}
\title{FUNCTION_TITLE}
\usage{
start_reactor(obj, silent = FALSE, processx_cleanup = TRUE)
}
\arguments{
\item{obj}{PARAM_DESCRIPTION}

\item{silent}{PARAM_DESCRIPTION}

\item{processx_cleanup}{PARAM_DESCRIPTION, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[processx]{process}}
}
