% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test_reactor}
\alias{test_reactor}
\title{Reactor testing Environment}
\usage{
test_reactor(
  expr,
  test_path = tempdir(),
  test_ip = "http://127.0.0.1",
  test_port = 6012,
  test_driver = firefox_driver(test_path = test_path),
  processx_args = runApp_args(test_port, test_path),
  processx_cleanup = TRUE
)
}
\arguments{
\item{expr}{expressions that drive the application}

\item{test_path}{character, path to run the test in. Default: tempdir()}

\item{test_ip}{character, IP address to run the App. Default: 'http://127.0.0.1'}

\item{test_port}{numeric, port to run the App. Default: 6012}

\item{test_driver}{\link[RSelenium:remoteDriver-class]{remoteDriver}, Default: firefox_driver(test_path = test_path)}

\item{processx_args}{list, arguments to pass to temporary session that serves the App. Default: runApp_args(test_ip, test_path)}

\item{processx_cleanup}{logical, cleanup the artifacts created by processx. Default: TRUE}
}
\value{
whereami counter data.frame
}
\description{
Environment which sets up and runs Shiny App tests.
}
\seealso{
\code{\link[processx]{process}}
}
