% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test_reactor}
\alias{test_reactor}
\title{FUNCTION_TITLE}
\usage{
test_reactor(expr, test_path = tempdir(), test_ip = 6012,
  test_driver = driver(test_path = test_path),
  processx_args = shinyAppFile_args(test_ip, test_path),
  processx_cleanup = TRUE)
}
\arguments{
\item{expr}{PARAM_DESCRIPTION}

\item{test_path}{PARAM_DESCRIPTION, Default: tempdir()}

\item{test_ip}{PARAM_DESCRIPTION, Default: 6012}

\item{test_driver}{PARAM_DESCRIPTION, Default: driver(test_path = test_path)}

\item{processx_args}{PARAM_DESCRIPTION, Default: shinyAppFile_args(test_ip, test_path)}

\item{processx_cleanup}{PARAM_DESCRIPTION, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[processx]{process}}
}
