% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test_whereami}
\alias{test_whereami}
\title{FUNCTION_TITLE}
\usage{
test_whereami(expr, test_path = tempdir(),
  test_driver = driver(test_path = test_path), test_ip = 6012)
}
\arguments{
\item{expr}{PARAM_DESCRIPTION}

\item{test_path}{PARAM_DESCRIPTION, Default: tempdir()}

\item{test_driver}{PARAM_DESCRIPTION, Default: driver(test_path = test_path)}

\item{test_ip}{PARAM_DESCRIPTION, Default: 6012}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[processx]{process}}
}
