% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test_whereami}
\alias{test_whereami}
\title{FUNCTION_TITLE}
\usage{
test_whereami(expr, dirpath = tempdir(), check = interactive(),
  verbose = interactive())
}
\arguments{
\item{expr}{PARAM_DESCRIPTION}

\item{dirpath}{PARAM_DESCRIPTION, Default: tempdir()}

\item{check}{PARAM_DESCRIPTION,  Default: interactive()}

\item{verbose}{PARAM_DESCRIPTION,  Default: interactive()}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[processx]{process}}
\code{\link[RSelenium]{rsDriver}}
}
