% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{text}
\alias{text}
\title{Generate an HTML text input}
\usage{
text(value = "", title = "")
}
\arguments{
\item{value}{initial value}

\item{title}{title of input}
}
\description{
Generate an HTML text input
}
\examples{
# Run this in a cell to display a text input
input <- text(value = "Default value", title = "Your title")

# In another cell, retrieve the value of the textbox:
as.character(input)

}
